change type row for wm_mashine

+++ frontend/models/dto/WmDto.php
@@ -20,7 +20,7 @@ class WmDto
     public function __construct($data)
     {
         if (array_key_exists('number_device', $data)) {
-            $this->number_device = (integer) $data['number_device'];
+            $this->number_device = (double) $data['number_device'];
         }
 
         if (array_key_exists('type_mashine', $data)) {
@@ -32,15 +32,15 @@ class WmDto
         }
 
         if (array_key_exists('bill_cash', $data)) {
-            $this->bill_cash = (integer) $data['bill_cash'];
+            $this->bill_cash = (float) $data['bill_cash'];
         }
 
         if (array_key_exists('door_position', $data)) {
-            $this->door_position = (integer) $data['door_position'];
+            $this->door_position = (double) $data['door_position'];
         }
 
         if (array_key_exists('door_block_led', $data)) {
-            $this->door_block_led = (integer) $data['door_block_led'];
+            $this->door_block_led = (double) $data['door_block_led'];
         }
 
         if (array_key_exists('current_status', $data)) {
